Bundle from bundle.js
Notes
=====
VegaObservatory must be global as it heads into the browser. Because
Browserify doesn't automatically but the package on the global
namespace, we have to put it there. This will allow us to access
VegaObservatory in the browser when bundling with Browserify, in
addition to using VegaObservatory as a properly accessible packge via
`require('vega-observatory')`